Rename df module to environment
