+++ src/environment/tests/pipe.test.ts
@@ -1,6 +1,6 @@
 import { assertEquals, describe, it, z } from '../../test-prelude.ts'
 import {
-  df,
+  environment,
   EnvironmentError,
   failure,
   InputError,
@@ -9,32 +9,22 @@ import {
 } from '../../index.ts'
 import type { Composable } from '../../index.ts'
 
-describe('sequence', () => {
-  it('should compose domain functions from left-to-right saving the results sequentially', async () => {
+describe('pipe', () => {
+  it('should compose domain functions from left-to-right', async () => {
     const a = withSchema(z.object({ id: z.number() }))(({ id }) => ({
       id: id + 2,
     }))
-    const b = withSchema(z.object({ id: z.number() }))(({ id }) => ({
-      result: id - 1,
-    }))
+    const b = withSchema(z.object({ id: z.number() }))(({ id }) => id - 1)
 
-    const c = df.sequence(a, b)
+    const c = environment.pipe(a, b)
     type _R = Expect<
       Equal<
         typeof c,
-        Composable<
-          (
-            input?: unknown,
-            environment?: unknown,
-          ) => [{ id: number }, { result: number }]
-        >
+        Composable<(input?: unknown, environment?: unknown) => number>
       >
     >
 
-    assertEquals(
-      await c({ id: 1 }),
-      success<[{ id: number }, { result: number }]>([{ id: 3 }, { result: 2 }]),
-    )
+    assertEquals(await c({ id: 1 }), success(2))
   })
 
   it('should use the same environment in all composed functions', async () => {
@@ -47,28 +37,17 @@ describe('sequence', () => {
     const b = withSchema(
       z.object({ inp: z.number() }),
       z.object({ env: z.number() }),
-    )(({ inp }, { env }) => ({ result: inp + env }))
+    )(({ inp }, { env }) => inp + env)
 
-    const c = df.sequence(a, b)
+    const c = environment.pipe(a, b)
     type _R = Expect<
       Equal<
         typeof c,
-        Composable<
-          (
-            input?: unknown,
-            environment?: unknown,
-          ) => [{ inp: number }, { result: number }]
-        >
+        Composable<(input?: unknown, environment?: unknown) => number>
       >
     >
 
-    assertEquals(
-      await c(undefined, { env: 1 }),
-      success<[{ inp: number }, { result: number }]>([
-        { inp: 3 },
-        { result: 4 },
-      ]),
-    )
+    assertEquals(await c(undefined, { env: 1 }), success(4))
   })
 
   it('should fail on the first environment parser failure', async () => {
@@ -84,13 +63,11 @@ describe('sequence', () => {
       envParser,
     )(({ inp }, { env }) => inp + env)
 
-    const c = df.sequence(a, b)
+    const c = environment.pipe(a, b)
     type _R = Expect<
       Equal<
         typeof c,
-        Composable<
-          (input?: unknown, environment?: unknown) => [{ inp: number }, number]
-        >
+        Composable<(input?: unknown, environment?: unknown) => number>
       >
     >
 
@@ -114,13 +91,11 @@ describe('sequence', () => {
       z.object({ env: z.number() }),
     )(({ inp }, { env }) => inp + env)
 
-    const c = df.sequence(a, b)
+    const c = environment.pipe(a, b)
     type _R = Expect<
       Equal<
         typeof c,
-        Composable<
-          (input?: unknown, environment?: unknown) => [{ inp: number }, number]
-        >
+        Composable<(input?: unknown, environment?: unknown) => number>
       >
     >
 
@@ -142,13 +117,11 @@ describe('sequence', () => {
       z.object({ env: z.number() }),
     )(({ inp }, { env }) => inp + env)
 
-    const c = df.sequence(a, b)
+    const c = environment.pipe(a, b)
     type _R = Expect<
       Equal<
         typeof c,
-        Composable<
-          (input?: unknown, environment?: unknown) => [{ inp: string }, number]
-        >
+        Composable<(input?: unknown, environment?: unknown) => number>
       >
     >
 
@@ -166,37 +139,17 @@ describe('sequence', () => {
       aBoolean: aString == '1',
     }))
     const c = withSchema(z.object({ aBoolean: z.boolean() }))(
-      ({ aBoolean }) => ({
-        anotherBoolean: !aBoolean,
-      }),
+      ({ aBoolean }) => !aBoolean,
     )
 
-    const d = df.sequence(a, b, c)
+    const d = environment.pipe(a, b, c)
     type _R = Expect<
       Equal<
         typeof d,
-        Composable<
-          (
-            input?: unknown,
-            environment?: unknown,
-          ) => [
-            { aString: string },
-            { aBoolean: boolean },
-            { anotherBoolean: boolean },
-          ]
-        >
+        Composable<(input?: unknown, environment?: unknown) => boolean>
       >
     >
 
-    assertEquals(
-      await d({ aNumber: 1 }),
-      success<
-        [
-          { aString: string },
-          { aBoolean: boolean },
-          { anotherBoolean: boolean },
-        ]
-      >([{ aString: '1' }, { aBoolean: true }, { anotherBoolean: false }]),
-    )
+    assertEquals(await d({ aNumber: 1 }), success(false))
   })
 })
