+++ src/index.ts
@@ -40,8 +40,8 @@ export type {
   UnpackData,
 } from './types.ts'
 
-// DOMAIN FUNCTIONS
-export * as df from './df/index.ts'
+// FUNCTIONS WITH ENVIRONMENT
+export * as environment from './environment/index.ts'
 
 // COMPAT MODULE
 export * as compat from './compat/index.ts'
